Rename texttosql to text2sql
