+++ comps/text2sql/src/chinook.sql
