+++ comps/text2sql/src/opea_text2sql_microservice.py
@@ -8,13 +8,13 @@ import sys
 from fastapi.exceptions import HTTPException
 
 from comps import CustomLogger, OpeaComponentController, opea_microservices, register_microservice
-from comps.texttosql.src.integrations.opea import Input, OpeaTextToSQL
+from comps.text2sql.src.integrations.opea import Input, OpeaText2SQL
 
 cur_path = pathlib.Path(__file__).parent.resolve()
 comps_path = os.path.join(cur_path, "../../../")
 sys.path.append(comps_path)
 
-logger = CustomLogger("texttosql")
+logger = CustomLogger("text2sql")
 logflag = os.getenv("LOGFLAG", False)
 
 try:
@@ -22,13 +22,13 @@ try:
     controller = OpeaComponentController()
 
     # Register components
-    texttosql_agent = OpeaTextToSQL(
-        name="TextToSQL",
-        description="TextToSQL Service",
+    text2sql_agent = OpeaText2SQL(
+        name="Text2SQL",
+        description="Text2SQL Service",
     )
 
     # Register components with the controller
-    controller.register(texttosql_agent)
+    controller.register(text2sql_agent)
 
     # Discover and activate a healthy component
     controller.discover_and_activate()
@@ -37,8 +37,8 @@ except Exception as e:
 
 
 @register_microservice(
-    name="opea_service@texttosql",
-    endpoint="/v1/texttosql",
+    name="opea_service@text2sql",
+    endpoint="/v1/text2sql",
     host="0.0.0.0",
     port=8080,
 )
@@ -46,7 +46,7 @@ async def execute_agent(input: Input):
     """Execute a SQL query from the input text.
 
     This function takes an Input object containing the input text and database connection information.
-    It uses the execute function from the texttosql module to execute the SQL query and returns the result.
+    It uses the execute function from the text2sql module to execute the SQL query and returns the result.
 
     Args:
         input (Input): An Input object with the input text and database connection information.
@@ -63,5 +63,5 @@ async def execute_agent(input: Input):
 
 
 if __name__ == "__main__":
-    logger.info("OPEA TextToSQL Microservice is starting...")
-    opea_microservices["opea_service@texttosql"].start()
+    logger.info("OPEA Text2SQL Microservice is starting...")
+    opea_microservices["opea_service@text2sql"].start()
