+++ comps/text2sql/src/requirements.txt
