+++ tests/text2sql/test_text2sql_opea.sh
@@ -18,7 +18,7 @@ export POSTGRES_DB=chinook
 
 function build_docker_images() {
     cd $WORKPATH
-    docker build --no-cache -t opea/texttosql:comps -f comps/texttosql/src/Dockerfile .
+    docker build --no-cache -t opea/text2sql:comps -f comps/text2sql/src/Dockerfile .
 }
 
 
@@ -51,22 +51,22 @@ check_tgi_connection() {
 
 function start_service() {
 
-    docker run --name test-texttosql-postgres --ipc=host -e POSTGRES_USER=${POSTGRES_USER} -e POSTGRES_HOST_AUTH_METHOD=trust -e POSTGRES_DB=${POSTGRES_DB} -e POSTGRES_PASSWORD=${POSTGRES_PASSWORD} -p 5442:5432 -d -v $WORKPATH/comps/texttosql/src/chinook.sql:/docker-entrypoint-initdb.d/chinook.sql postgres:latest
+    docker run --name test-text2sql-postgres --ipc=host -e POSTGRES_USER=${POSTGRES_USER} -e POSTGRES_HOST_AUTH_METHOD=trust -e POSTGRES_DB=${POSTGRES_DB} -e POSTGRES_PASSWORD=${POSTGRES_PASSWORD} -p 5442:5432 -d -v $WORKPATH/comps/text2sql/src/chinook.sql:/docker-entrypoint-initdb.d/chinook.sql postgres:latest
 
-    docker run -d --name="test-texttosql-tgi-endpoint" --ipc=host -p $tgi_port:80 -v ./data:/data --shm-size 1g -e HUGGINGFACEHUB_API_TOKEN=${HUGGINGFACEHUB_API_TOKEN} -e HF_TOKEN=${HF_TOKEN} -e model=${model} ghcr.io/huggingface/text-generation-inference:2.1.0 --model-id $model
+    docker run -d --name="test-text2sql-tgi-endpoint" --ipc=host -p $tgi_port:80 -v ./data:/data --shm-size 1g -e HUGGINGFACEHUB_API_TOKEN=${HUGGINGFACEHUB_API_TOKEN} -e HF_TOKEN=${HF_TOKEN} -e model=${model} ghcr.io/huggingface/text-generation-inference:2.1.0 --model-id $model
 
     export TGI_LLM_ENDPOINT="http://${ip_address}:${tgi_port}"
-    texttosql_port=9090
+    text2sql_port=9090
     unset http_proxy
-    docker run -d --name="test-texttosql-server" --ipc=host -p ${texttosql_port}:8080 --ipc=host -e http_proxy=$http_proxy -e https_proxy=$https_proxy -e TGI_LLM_ENDPOINT=$TGI_LLM_ENDPOINT opea/texttosql:comps
+    docker run -d --name="test-text2sql-server" --ipc=host -p ${text2sql_port}:8080 --ipc=host -e http_proxy=$http_proxy -e https_proxy=$https_proxy -e TGI_LLM_ENDPOINT=$TGI_LLM_ENDPOINT opea/text2sql:comps
 
     # check whether tgi is fully ready
     check_tgi_connection "${TGI_LLM_ENDPOINT}/health"
 }
 
 function validate_microservice() {
-    texttosql_port=9090
-    result=$(http_proxy="" curl http://${ip_address}:${texttosql_port}/v1/texttosql\
+    text2sql_port=9090
+    result=$(http_proxy="" curl http://${ip_address}:${text2sql_port}/v1/text2sql\
         -X POST \
         -d '{"input_text": "Find the total number of Albums.","conn_str": {"user": "'${POSTGRES_USER}'","password": "'${POSTGRES_PASSWORD}'","host": "'${ip_address}'", "port": "5442", "database": "'${POSTGRES_DB}'" }}' \
         -H 'Content-Type: application/json')
@@ -76,15 +76,15 @@ function validate_microservice() {
         echo "Result correct."
     else
         echo "Result wrong. Received was $result"
-        docker logs test-texttosql-server > ${LOG_PATH}/texttosql.log
-        docker logs test-texttosql-tgi-endpoint > ${LOG_PATH}/tgi.log
+        docker logs test-text2sql-server > ${LOG_PATH}/text2sql.log
+        docker logs test-text2sql-tgi-endpoint > ${LOG_PATH}/tgi.log
         exit 1
     fi
 
 }
 
 function stop_docker() {
-    cid=$(docker ps -aq --filter "name=test-texttosql*")
+    cid=$(docker ps -aq --filter "name=test-text2sql*")
     if [[ ! -z "$cid" ]]; then docker stop $cid && docker rm $cid && sleep 1s; fi
 }
